requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
future==0.18.2
idna==2.10
multidict==5.1.0
oauthlib==3.1.1
parallelmediadownloader==0.1.0
python-twitter==3.5
requests==2.25.1
requests-oauthlib==1.3.0
twitter-dl==0.2.0
typing-extensions==3.10.0.0
urllib3==1.26.6
yarl==1.6.3
you-get==0.4.1527
